pyproject.toml
[project]
name = "kaggle-mcp-server"
version = "0.1.0"
description = "A Model Context Protocol server for Kaggle API integration"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.10.0",
    "kaggle>=1.7.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kaggle_mcp_server"]
